langchain-example/pyproject.toml
[project]
name = "langchain-example"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "yuanbohan", email = "yuanbo.han@gmail.com" }
]
dependencies = [
    "openai>=0.28.0",
    "langchain>=0.0.27",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-api>=1.20.0",
    "Flask>=2.3.3",
    "prometheus-client>=0.17.1",
    "opentelemetry-exporter-prometheus>=1.12.0rc1",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "chromadb>=0.4.10",
    "tiktoken>=0.5.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "greptime-llm-langchain-instrument @ file:///home/yuanbo/workspace/github.com/greptime-llm-instrument/langchain-example/../langchain",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.rye.scripts]
app = { cmd = "python src/langchain_example/app.py" }
